Capricornus NFT
Capricornus is a faint zodiac constellation located in the southern sky. Its name means “the goat” in Latin. The constellation represents a sea-goat, a mythical creature associated with the god Enki in Babylonian mythology and later with the Greek deity Pan. The constellation is represented by the symbol ♑. 

Capricorns are the hardest workers of the zodiac and love nothing more than getting ahead in life. They are ambitious, determined, materialistic, and strong. They will keep going when others would’ve given up ten miles back. This makes them great partners in life, as well as friends or collaborators.